% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_class.R
\name{col_class}
\alias{col_class}
\title{Returns all columns and classes matching specified class}
\usage{
col_class(x, find_class = NULL)
}
\arguments{
\item{x}{a data.frame object with names columns}

\item{find_class}{a single or vector of character value(s) of column class
types}
}
\description{
The \code{col_class} function takes two arguments, \code{x}, a data.frame
object, and \code{find_class}, a single or vector of character value(s) of
the class type desired.
}
\examples{
x <- data.frame(
  col_a = as.factor(c("a", "b", "c"))
  , col_b = as.integer(c(1, 2, 3))
  , col_c = as.factor(c("apple", "banana", "citrus"))
  , col_d = as.character(c("one", "two", "three"))
)
col_class(x, "factor")
col_class(x, c("integer", "character"))
}
