% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_class.R
\name{column_class}
\alias{column_class}
\title{Returns all columns and classes matching specified class}
\usage{
column_class(.data = NULL, find_class = NULL)
}
\arguments{
\item{.data}{data.frame; a \bold{data.frame} with named columns}

\item{find_class}{string or string vector; a single or vector of string
value(s) of column class types}
}
\description{
The \code{column_class} function takes two arguments,
  \code{.data}, a data.frame object, and \code{find_class}, a single or
  vector of character value(s) of the class type desired.
}
\examples{
x <- data.frame(
  col_a = as.factor(c("a", "b", "c"))
  , col_b = as.integer(c(1, 2, 3))
  , col_c = as.factor(c("apple", "banana", "citrus"))
  , col_d = as.character(c("one", "two", "three"))
)
column_class(x, "factor")
column_class(x, c("integer", "character"))
}
