% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_experiment.R
\name{generate_experiment}
\alias{generate_experiment}
\title{Generate a Data Frame with Basic Experimental Data}
\usage{
generate_experiment(n = 1000, threshold = 0.01)
}
\arguments{
\item{n}{a numeric, integer value indicating the number of observations in
the experiment}

\item{threshold}{a numeric, float value indicating the threshold cutoff
where a value less than or equal to the threshold is considered a success
(TRUE) and a value greater than the cutoff is a failure (FALSE)}
}
\description{
\code{generate_experiment} takes two arguments, n is the number
  of experimental observations, and threshold is the success/failure
  threshold level
}
\keyword{experiment}
