% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_number.R
\name{generate_random_number}
\alias{generate_random_number}
\title{Random Number Generator}
\usage{
generate_random_number(n = NULL, min = NULL, max = NULL,
  type = "integer", digits = NULL)
}
\arguments{
\item{n}{a numeric argument, specifies the number of random numbers the user
would like returned}

\item{min}{a numeric argument, specifies the minimum number in the range to
draw from}

\item{max}{a numeric argument, specifies the maximum number in the range to
draw from}

\item{type}{a character argument, must be either \code{"integer"} or
\code{"float"}, specifies the format of the desired, numeric output as
either an \code{integer} or \code{float} value}

\item{digits}{\code{optional} a number value, to accompany
\code{type = "float"} and specify the number of decimal digits to include}
}
\description{
\code{generate_random_number} draws a random number(s) from a range either as an integer or a
  float with a specified number of decimal digits
}
\examples{
generate_random_number(n = 1, min = 1, max = 10, type = "integer")

generate_random_number(n = 2, min = 1, max = 10, type = "float", digits = 3)

generate_random_number(n = 3, min = 1, max = 10, type = "float", digits = "5")
}
