% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_files_batch.R
\name{import_batch_files}
\alias{import_batch_files}
\title{Import a batch of files}
\usage{
import_batch_files(p = NULL, f = NULL, ext = NULL, ...)
}
\arguments{
\item{p}{A string of the directory path containing desired files.}

\item{f}{A function name used for importing of individual files.}

\item{ext}{Optional. A character string \itemize{
\item \code{".csv"}
\item \code{".tsv"}
\item \code{".xlsx"}
}}
}
\value{
array with dimensions as sub-frame rows, sub-frame columns, and
 and sub-frame index.
}
\description{
To import one or more files into a single df object. The files
  should all be housed in the same directory, the path to which is supplied
  as the sole argument.
}
