% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_digits.R
\name{number_digits}
\alias{number_digits}
\title{Count Number of Decimal Digits Function}
\usage{
number_digits(x)
}
\arguments{
\item{x}{scalar object.}
}
\description{
\code{number_digits} takes one argument, an object x, that will
  be evaluated for the number of decimal digits it contains.
}
\keyword{decimal}
\keyword{digits}
