% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_num.R
\name{rand_num}
\alias{rand_num}
\title{Random Number Generator}
\usage{
rand_num(n = NULL, min = NULL, max = NULL, type = "integer",
  digits = NULL)
}
\arguments{
\item{n}{a numeric argument, specifies the number of random numbers the user
would like returned}

\item{min}{a numeric argument, specifies the minimum number in the range to
draw from}

\item{max}{a numeric argument, specifies the maximum number in the range to
draw from}

\item{type}{a character argument, must be either \code{"integer"} or
\code{"float"}, specifies the format of the desired, numeric output as
either an \code{integer} or \code{float} value}

\item{digits}{\code{optional} a number value, to accompany
\code{type = "float"} and specify the number of decimal digits to include}
}
\description{
\code{rand_num} draws a random number(s) from a range either as an integer or a
  float with a specified number of decimal digits
}
\examples{
rand_num(min = 1, max = 10, type = "integer")

rand_num(min = 1, max = 10, type = "float", digits = 3)

rand_num(min = 1, max = 10, type = "float", digits = "5")
}
