% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randnum.R
\name{randnum}
\alias{randnum}
\title{Random Number Generator}
\usage{
randnum(min = 1, max = 1000, type = "integer", digits = NA)
}
\arguments{
\item{min}{a numeric argument, specifies the minimum number in the range to
draw from}

\item{max}{a numeric argument, specifies the maximum number in the range to
draw from}

\item{type}{a character argument, must be either \code{"integer"} or
\code{"float"}, specifies the format of the desired, numeric output as
either an \code{integer} or \code{float} value}

\item{digits}{\code{optional} a number value, to accompany
\code{type = "float"} and specify the number of decimal digits to include}
}
\description{
\code{randnum} draws a random number from a range either as an integer or a
  float with a specified number of decimal digits
}
\examples{
randnum(min = 1, max = 10, type = "integer")

randnum(min = 1, max = 10, type = "float", digits = 3)

randnum(min = 1, max = 10, type = "float", digits = "5")
}
