% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_sample.R
\name{rows_sample}
\alias{rows_sample}
\title{Sample Random Rows from Data Frame}
\usage{
rows_sample(df, n, replace = FALSE)
}
\arguments{
\item{df}{source data frame to be sampled from}

\item{n}{sample size}

\item{replace}{shoould the smapling method use repalcement? \code{TRUE} if
yes, \code{FALSE} (by default) if no}
}
\description{
It is common to want a random sampling of data from a large data set for
  analysis of various forms. The \code{rows_sample} function takes
  arguments needed to return a random sample of a data frame. Providing the
  source data frame, a sample size, and whether or not replacement is
  allowed in the sampling is all that is required for returning the sample.
  Row names (number typically) are returned in the sample making it easy to
  identify which observations are selected into the sample.
}
\examples{
iris_samp <- rows_sample(iris, 5, replace = FALSE)
}
\keyword{random}
\keyword{row}
\keyword{sample}
