% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_columns.R
\name{sort_cols}
\alias{sort_cols}
\title{Dataframe Column Sorting Function
\code{sort_cols} takes two arguments, the name of a dataframe and the order
  (ascending or descending), and will then provide the dataframe in the
  column sorting as specified.}
\usage{
sort_cols(x = NULL, order = "asc")
}
\arguments{
\item{x}{a dataframe object.}

\item{order}{asc or desc for ascending or descending ordering respectively.}
}
\description{
Dataframe Column Sorting Function
\code{sort_cols} takes two arguments, the name of a dataframe and the order
  (ascending or descending), and will then provide the dataframe in the
  column sorting as specified.
}
\keyword{columns}
\keyword{dataframe}
\keyword{sort}
