% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_frame.R
\name{subset_frame}
\alias{subset_frame}
\title{A sub-framing function}
\usage{
subset_frame(df = NULL, num_subs = NULL)
}
\arguments{
\item{df}{data frame with any number of columns or rows to be subsetted to
equivalently sized sub-frames.}

\item{num_subs}{single, numeric value of the number of sub-frames desired.
Default is \code{NULL} in which case, if no value is entered, the frame
will be divided into 10 sub-frames.}
}
\value{
array with dimensions as sub-frame rows, sub-frame columns, and
 and sub-frame index.
}
\description{
The \code{"subset_frame"} function is a subsetting function intended to
  produce equivalently size subsets from a single data frame. This is done
  by evaluating the number of observations in the original data set and
  calculating how many rows should be in each subset for a given nubmer of
  subsets.
The purpose behind \code{"subset_frame"} is to break apart large data frames
  into smaller ones for the sake of efficiency. This could be useful for
  test sets for code or method testing, etc.
}
\examples{
# Data frame of 100 random observations from normal and uniform
df <- data.frame("X" = rnorm(100, 0, 1), "Y" = runif(100, 0, 1))

# Sub-frame to 2 data frames
subset_frame(df, 2)

# Sub-frame with default sub-framing
subset_frame(df)
}
