% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_batch_files.R
\name{import_batch_files}
\alias{import_batch_files}
\title{Import a batch of files}
\usage{
import_batch_files(.path = NULL, .fun = NULL, .ext = NULL, ...)
}
\arguments{
\item{.path}{string; the \bold{path} to the directory containing desired files.}

\item{.fun}{string; the \bold{function name} used for importing files.}

\item{.ext}{\code{(optional)} string; the file \bold{extension} to indicate
type of external data file to be imported:\itemize{
\item \code{".csv"}
\item \code{".tsv"}
\item \code{".xlsx"}
}}

\item{...}{\code{(optional)} additional arguments specific to \code{f}.}
}
\value{
An array with dimensions as sub-frame rows, sub-frame columns, and
 and sub-frame index.
}
\description{
To import one or more files into a single df object. The files
  should all be housed in the same directory, the path to which is supplied
  as the sole argument.
}
