% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_compliment.R
\name{sample_compliment}
\alias{sample_compliment}
\title{A Sampling Compliment Function}
\usage{
sample_compliment(orig, samp)
}
\arguments{
\item{orig}{a data.frame; the original/complete data set being sampled}

\item{samp}{a data.frame; a subset, randomly sampled from the \code{orig}}
}
\description{
Designed to work in tandem with \code{sample_rows} to return
  a complimentary data set to a sample from an original set.
}
\examples{
iris_test <- sample_rows(
  df = iris
  , n = floor(0.2 * nrow(iris))
  , s = 1
  , replace = FALSE
)
iris_train <- sample_compliment(orig = iris, samp = iris_test)
}
