% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_rows.R
\name{sample_rows}
\alias{sample_rows}
\title{Sample Random Rows from Data Frame}
\usage{
sample_rows(.data = NULL, n, .seed = 1, replace = FALSE)
}
\arguments{
\item{.data}{an object; the source \bold{data.frame} to be sampled from}

\item{n}{a numeric value; sample size}

\item{.seed}{\code{(optional)} a numeric value; the \bold{seed} for randomly
generating numbers.}

\item{replace}{a logical value; should the sampling method use replacement?
\code{TRUE} if yes, \code{FALSE} (by default) if no}
}
\description{
It is common to want a random sampling of data from a large
  data set for analysis of various forms. The \code{sample_rows} function
  takes arguments needed to return a random sample of a data frame.
  Providing the source data frame, a sample size, and whether or not
  replacement is allowed in the sampling is all that is required for
  returning the sample. Row names (number typically) are returned in the
  sample making it easy to identify which observations are selected into the
  sample.
}
\examples{
iris_samp <- sample_rows(iris, 5, replace = FALSE)
}
\keyword{random}
\keyword{row}
\keyword{sample}
