% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_columns.R
\name{sort_columns}
\alias{sort_columns}
\title{Dataframe Column Sorting Function
\code{sort_columns} takes two arguments, the name of a dataframe and the
  order (ascending or descending), and will then provide the dataframe in
  the column sorting as specified.}
\usage{
sort_columns(x = NULL, order = "asc", verbose = FALSE)
}
\arguments{
\item{x}{dataframe object}

\item{order}{character string; if \code{"asc"}, the function sorts in
ascending order; if \code{"desc"}, the function sorts in descending order.}

\item{verbose}{logical; if \code{TRUE}, the function prints messages about
objects being sorted.}
}
\description{
Dataframe Column Sorting Function
\code{sort_columns} takes two arguments, the name of a dataframe and the
  order (ascending or descending), and will then provide the dataframe in
  the column sorting as specified.
}
\keyword{columns}
\keyword{sort}
